NP-2474 adapt all the test to the new organization
