+++ internal/pkg/provider/organization/provider_test.go
@@ -18,11 +18,22 @@ package organization
 
 import (
 	"fmt"
+	"github.com/google/uuid"
 	"github.com/nalej/system-model/internal/pkg/entities"
-	"github.com/nalej/system-model/internal/pkg/utils"
 	"github.com/onsi/ginkgo"
 	"github.com/onsi/gomega"
 )
+
+func CreateOrganization() *entities.Organization {
+	return entities.NewOrganization(
+		fmt.Sprintf(fmt.Sprintf("org-%s", uuid.New().String())),
+		"Nalej Test Address",
+		"City Test",
+		"State Test",
+		"U.S.A",
+		"XXX",
+		"Photo")
+}
 
 func RunTest(provider Provider) {
 
@@ -33,7 +44,7 @@ func RunTest(provider Provider) {
 	// Add and organization
 	ginkgo.It("Should be able to add a organization", func() {
 
-		org := utils.CreateOrganization()
+		org := CreateOrganization()
 		err := provider.Add(*org)
 		gomega.Expect(err).To(gomega.Succeed())
 
@@ -42,7 +53,7 @@ func RunTest(provider Provider) {
 	// Get Organization
 	ginkgo.It("Should be able to get a organization", func() {
 
-		org := utils.CreateOrganization()
+		org := CreateOrganization()
 
 		err := provider.Add(*org)
 		gomega.Expect(err).To(gomega.Succeed())
@@ -57,8 +68,8 @@ func RunTest(provider Provider) {
 	// List Organization
 	ginkgo.It("Should be able to list a organization", func() {
 
-		org := utils.CreateOrganization()
-		org1 := utils.CreateOrganization()
+		org := CreateOrganization()
+		org1 := CreateOrganization()
 
 		err := provider.Add(*org)
 		gomega.Expect(err).To(gomega.Succeed())
@@ -93,7 +104,7 @@ func RunTest(provider Provider) {
 	// Exists Organization
 	ginkgo.It("Should be able to find a organization", func() {
 
-		org := utils.CreateOrganization()
+		org := CreateOrganization()
 
 		err := provider.Add(*org)
 		gomega.Expect(err).To(gomega.Succeed())
@@ -111,7 +122,7 @@ func RunTest(provider Provider) {
 	})
 	ginkgo.It("Should be able to find a organization by name", func() {
 
-		org := utils.CreateOrganization()
+		org := CreateOrganization()
 
 		err := provider.Add(*org)
 		gomega.Expect(err).To(gomega.Succeed())
@@ -129,7 +140,7 @@ func RunTest(provider Provider) {
 	})
 
 	ginkgo.It("Should be able to update an organization", func() {
-		org := utils.CreateOrganization()
+		org := CreateOrganization()
 		// add the organization
 		err := provider.Add(*org)
 		gomega.Expect(err).To(gomega.Succeed())
@@ -148,12 +159,12 @@ func RunTest(provider Provider) {
 
 	})
 	ginkgo.It("Should not be able to update an organization if exists other one with the new name", func() {
-		org1 := utils.CreateOrganization()
+		org1 := CreateOrganization()
 		// add the organization
 		err := provider.Add(*org1)
 		gomega.Expect(err).To(gomega.Succeed())
 
-		org2 := utils.CreateOrganization()
+		org2 := CreateOrganization()
 		// add the organization
 		err = provider.Add(*org2)
 		gomega.Expect(err).To(gomega.Succeed())
@@ -167,7 +178,7 @@ func RunTest(provider Provider) {
 
 	})
 	ginkgo.It("Should not be able to update a non existing organization", func() {
-		org := utils.CreateOrganization()
+		org := CreateOrganization()
 		err := provider.Update(*org)
 		gomega.Expect(err).NotTo(gomega.Succeed())
 	})
