+++ internal/pkg/server/eic/handler_test.go
@@ -85,7 +85,7 @@ var _ = ginkgo.Describe("Asset service", func() {
 			organizationProvider.(*orgProvider.MockupOrganizationProvider).Clear()
 			controllerProvider.Clear()
 			// Initial data
-			targetOrganization = testhelpers.CreateOrganization(organizationProvider)
+			targetOrganization = testhelpers.AddOrganization(organizationProvider)
 		})
 	})
