+++ internal/pkg/server/organization/handler_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/nalej/grpc-organization-go"
 	"github.com/nalej/grpc-utils/pkg/test"
 	"github.com/nalej/system-model/internal/pkg/provider/organization"
-	"github.com/nalej/system-model/internal/pkg/utils"
+	"github.com/nalej/system-model/internal/pkg/server/testhelpers"
 	"github.com/onsi/ginkgo"
 	"github.com/onsi/gomega"
 	"google.golang.org/grpc"
@@ -43,13 +43,14 @@ var _ = ginkgo.Describe("Organization service", func() {
 	ginkgo.BeforeSuite(func() {
 		listener = test.GetDefaultListener()
 		server = grpc.NewServer()
-		test.LaunchServer(server, listener)
 
 		// Register the service
 		orgProvider = organization.NewMockupOrganizationProvider()
 		manager := NewManager(orgProvider)
 		handler := NewHandler(manager)
 		grpc_organization_go.RegisterOrganizationsServer(server, handler)
+
+		test.LaunchServer(server, listener)
 
 		conn, err := test.GetConn(*listener)
 		gomega.Expect(err).Should(gomega.Succeed())
@@ -67,7 +68,7 @@ var _ = ginkgo.Describe("Organization service", func() {
 
 	ginkgo.Context("adding organization", func() {
 		ginkgo.It("should support adding a new organization", func() {
-			toAdd := utils.CreateAddOrganizationRequest()
+			toAdd := testhelpers.CreateAddOrganizationRequest()
 			org, err := client.AddOrganization(context.Background(), toAdd)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org).ShouldNot(gomega.BeNil())
@@ -89,7 +90,7 @@ var _ = ginkgo.Describe("Organization service", func() {
 		})
 
 		ginkgo.It("should fail if the organization name already exists", func() {
-			toAdd := utils.CreateAddOrganizationRequest()
+			toAdd := testhelpers.CreateAddOrganizationRequest()
 			org, err := client.AddOrganization(context.Background(), toAdd)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org).ShouldNot(gomega.BeNil())
@@ -103,7 +104,7 @@ var _ = ginkgo.Describe("Organization service", func() {
 
 	ginkgo.Context("retrieve organization", func() {
 		ginkgo.It("should work on existing organization", func() {
-			toAdd := utils.CreateAddOrganizationRequest()
+			toAdd := testhelpers.CreateAddOrganizationRequest()
 			org, err := client.AddOrganization(context.Background(), toAdd)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org).ShouldNot(gomega.BeNil())
@@ -118,12 +119,12 @@ var _ = ginkgo.Describe("Organization service", func() {
 		})
 
 		ginkgo.It("should recover a list of organizations", func() {
-			toAdd := utils.CreateAddOrganizationRequest()
+			toAdd := testhelpers.CreateAddOrganizationRequest()
 			org, err := client.AddOrganization(context.Background(), toAdd)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org).ShouldNot(gomega.BeNil())
 
-			toAdd =  utils.CreateAddOrganizationRequest()
+			toAdd =  testhelpers.CreateAddOrganizationRequest()
 			org, err = client.AddOrganization(context.Background(), toAdd)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org).ShouldNot(gomega.BeNil())
@@ -161,12 +162,12 @@ var _ = ginkgo.Describe("Organization service", func() {
 
 	ginkgo.Context("update organization", func() {
 		ginkgo.It("should support updating an existing organization", func() {
-			toAdd := utils.CreateAddOrganizationRequest()
+			toAdd := testhelpers.CreateAddOrganizationRequest()
 			org, err := client.AddOrganization(context.Background(), toAdd)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org).ShouldNot(gomega.BeNil())
 
-			toUpdate := utils.CreateUpdateOrganizationRequest(org.OrganizationId, false, "")
+			toUpdate := testhelpers.CreateUpdateOrganizationRequest(org.OrganizationId, false, "")
 			success, err := client.UpdateOrganization(context.Background(), toUpdate)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(success).NotTo(gomega.BeNil())
@@ -178,35 +179,35 @@ var _ = ginkgo.Describe("Organization service", func() {
 		})
 
 		ginkgo.It("should fail on non-existing organization", func() {
-			toUpdate := utils.CreateUpdateOrganizationRequest(uuid.New().String(), false, "")
+			toUpdate := testhelpers.CreateUpdateOrganizationRequest(uuid.New().String(), false, "")
 			success, err := client.UpdateOrganization(context.Background(), toUpdate)
 			gomega.Expect(err).ShouldNot(gomega.Succeed())
 			gomega.Expect(success).To(gomega.BeNil())
 		})
 
 		ginkgo.It("should fail when removing the name of an organization", func() {
-			toAdd := utils.CreateAddOrganizationRequest()
+			toAdd := testhelpers.CreateAddOrganizationRequest()
 			org, err := client.AddOrganization(context.Background(), toAdd)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org).ShouldNot(gomega.BeNil())
 
-			toUpdate := utils.CreateUpdateOrganizationRequest(org.OrganizationId, true, "")
+			toUpdate := testhelpers.CreateUpdateOrganizationRequest(org.OrganizationId, true, "")
 			success, err := client.UpdateOrganization(context.Background(), toUpdate)
 			gomega.Expect(err).ShouldNot(gomega.Succeed())
 			gomega.Expect(success).To(gomega.BeNil())
 		})
 		ginkgo.It("Should fail when updating the name of an organization if there is another with that name", func() {
-			toAdd1 := utils.CreateAddOrganizationRequest()
+			toAdd1 := testhelpers.CreateAddOrganizationRequest()
 			org1, err := client.AddOrganization(context.Background(), toAdd1)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org1).ShouldNot(gomega.BeNil())
 
-			toAdd2 := utils.CreateAddOrganizationRequest()
+			toAdd2 := testhelpers.CreateAddOrganizationRequest()
 			org2, err := client.AddOrganization(context.Background(), toAdd2)
 			gomega.Expect(err).Should(gomega.Succeed())
 			gomega.Expect(org2).ShouldNot(gomega.BeNil())
 
-			toUpdate := utils.CreateUpdateOrganizationRequest(org1.OrganizationId, true, org2.Name)
+			toUpdate := testhelpers.CreateUpdateOrganizationRequest(org1.OrganizationId, true, org2.Name)
 			success, err := client.UpdateOrganization(context.Background(), toUpdate)
 			gomega.Expect(err).ShouldNot(gomega.Succeed())
 			gomega.Expect(success).To(gomega.BeNil())
